Workaround long ID3v2 tags problems
There was some strange code for dealing with ID3v2 tags that were longer
than 128 bytes. With Unicode tags this is only 64 characters so can happen
quite easily.

Add a test case for the problem and then just make the code cope with tags
up to 64K. It will break beyond that still as far as I can see but I've no
idea what the code was actually trying to do.