mp3,tests: Cope with files larger than ~71.5MiB
Multiplying the integer file size by FRAME_COUNT will overflow if the file
size is greater than MAX_INT/FRAME_COUNT =~= 71.5MiB. Let's promote the
size to a double before multiplying to avoid that.

Add a new test program to confirm that the fix is effective. This test
program requires a large file, so let's generate it by repeating a single
MP3 frame.